sonic-frr: Add encoding function for full nexthop group info
The logic of encoding is aligned with current NH operation
but with full information.
For singleton nexthop group, we encode its explicit information value.
For multi-path nexthop group, we encode the compressed id array
of depends and dependents tree, including the recursive nodes.
Besides, hash value is encoded for both above.

Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>